requirements.txt
python-binance==1.0.19
ccxt==4.1.77
fastapi>=0.111.0,<0.112.0
uvicorn==0.24.0
python-multipart>=0.0.7,<0.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets>=10.3,<11.0
aiohttp>=3.9.5,<4.0.0
pandas==2.1.4
numpy==1.25.2
pydantic>=2.7.0,<2.10.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.33.3
cryptography>=3.4.8
httpx>=0.23.0,<0.24.0
requests>=2.32.1,<3.0.0
pyOpenSSL>=23.3.0
nest-asyncio>=1.5.8
